refactor: Change project name / packages to PopularMovies for generality and incrementation from stage one to two
