+++ .idea/dictionaries/alexl.xml
@@ -2,6 +2,7 @@
   <dictionary name="alexl">
     <words>
       <w>krinotech</w>
+      <w>popularmovies</w>
       <w>popularmoviesstageone</w>
     </words>
   </dictionary>
