+++ app/src/main/java/com/krinotech/popularmovies/DetailsActivity.java
@@ -1,4 +1,4 @@
-package com.krinotech.popularmoviesstageone;
+package com.krinotech.popularmovies;
 
 import androidx.appcompat.app.AppCompatActivity;
