+++ app/src/test/java/com/krinotech/popularmovies/util/MovieJsonUtilTest.java
@@ -1,4 +1,4 @@
-package com.krinotech.popularmoviesstageone.util;
+package com.krinotech.popularmovies.util;
 
 import org.junit.Test;
